Add UserInput class and move ReadInput and checkInputAccuracy function from Main method to that class
